@@ -1,14 +1,12 @@
 import 'dart:convert';
-import 'dart:io';
 
 import 'package:HackRU/models/models.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:path_provider/path_provider.dart';
 
 Future<SharedPreferences> _prefs = SharedPreferences.getInstance();
 
 Future<void> deleteCredentials() async {
-  final SharedPreferences prefs = await _prefs;
+  final prefs = await _prefs;
   await prefs.remove('AUTH_TOKEN');
   await prefs.remove('EMAIL');
   await Future.delayed(Duration(seconds: 1));
@@ -16,7 +14,7 @@ Future<void> deleteCredentials() async {
 }
 
 Future<void> persistCredentials(String kAuthToken, String kEmail) async {
-  final SharedPreferences prefs = await _prefs;
+  final prefs = await _prefs;
   await prefs.setString('AUTH_TOKEN', kAuthToken);
   await prefs.setString('EMAIL', kEmail);
   await Future.delayed(Duration(seconds: 1));
@@ -24,7 +22,7 @@ Future<void> persistCredentials(String kAuthToken, String kEmail) async {
 }
 
 Future<bool> hasCredentials() async {
-  final SharedPreferences prefs = await _prefs;
+  final prefs = await _prefs;
   if (prefs.containsKey('AUTH_TOKEN')) {
     return true;
   }
@@ -33,7 +31,7 @@ Future<bool> hasCredentials() async {
 }
 
 Future<String> getAuthToken() async {
-  final SharedPreferences prefs = await _prefs;
+  final prefs = await _prefs;
   if (prefs.containsKey('AUTH_TOKEN')) {
     return prefs.getString('AUTH_TOKEN');
   }
@@ -42,7 +40,7 @@ Future<String> getAuthToken() async {
 }
 
 Future<String> getEmail() async {
-  final SharedPreferences prefs = await _prefs;
+  final prefs = await _prefs;
   if (prefs.containsKey('EMAIL')) {
     return prefs.getString('EMAIL');
   }
@@ -50,42 +48,31 @@ Future<String> getEmail() async {
   return '';
 }
 
+Future<void> persistSlackAnnouncements(List<Announcement> slacks) async {
+  var tsNow = (DateTime.now().millisecondsSinceEpoch ~/ 1000).toString();
 
-///================= STORING SLACK MESSAGES =================
-
-Future<String> _appPath() async {
-  return (await getApplicationDocumentsDirectory()).path;
+  final prefs = await _prefs;
+  var slacksMsgs = slacks.length > 1
+      ? '[' + slacks.map((slack) => slack.toString()).toList().join(',') + ']'
+      : '[{"text": "Error: Unable to retrieve messages!", "ts": "$tsNow"}]';
+  await prefs.setString('SLACK_ANNOUNCEMENTS', slacksMsgs);
+  await Future.delayed(Duration(seconds: 1));
 }
 
-Future<File> storedFile(String name) async {
-  var path = (await _appPath()) + "/" + name;
-  var f = File(path);
-  if (!(await f.exists())) {
-    await f.create();
+Future<List<Announcement>> getSlackAnnouncements() async {
+  final prefs = await _prefs;
+  if (prefs.containsKey('SLACK_ANNOUNCEMENTS')) {
+    var decoded = json.decode(prefs.getString('SLACK_ANNOUNCEMENTS'));
+    return decoded
+        .map<Announcement>((slack) => Announcement.fromJson(slack))
+        .toList();
   }
-  return f;
+  return null;
 }
 
-Future<File> storedSlacksFile() => storedFile("slacks.json");
-
-Future<List<Announcement>> getStoredSlacks() async {
-  var slacksFile = await storedSlacksFile();
-  var contents = await slacksFile.readAsString();
-  if (contents == "") {
-    return null;
-  }
-  List<dynamic> decoded = json.decode(contents);
-  return decoded.map((slack) => Announcement.fromJson(slack)).toList();
-}
-
-setStoredSlacks(List<Announcement> slacks) async {
-  var slacksFile = await storedSlacksFile();
-  var slacksString =
-      "[" + slacks.map((slack) => slack.toString()).toList().join(",") + "]";
-  await slacksFile.writeAsString(slacksString);
-}
-
-Future<void> deleteStoredSlacks() async {
-  var slacksFile = await storedSlacksFile();
-  await slacksFile.delete();
+Future<void> deleteSlackAnnouncements() async {
+  final prefs = await _prefs;
+  await prefs.remove('SLACK_ANNOUNCEMENTS');
+  await Future.delayed(Duration(seconds: 1));
+  return;
 }
