@@ -9,10 +9,9 @@ import 'package:HackRU/ui/pages/qr_scanner/QRScanner.dart';
 import 'package:flare_flutter/flare_actor.dart';
 import 'package:flutter/material.dart';
 import 'package:groovin_material_icons/groovin_material_icons.dart';
-import 'package:hidden_drawer_menu/hidden_drawer/hidden_drawer_menu.dart';
-import 'package:hidden_drawer_menu/hidden_drawer/screen_hidden_drawer.dart';
-import 'package:hidden_drawer_menu/menu/item_hidden_menu.dart';
 import 'package:HackRU/ui/widgets/custom_hidden_drawer_menu.dart';
+import 'package:hidden_drawer_menu/hidden_drawer_menu.dart';
+
 import '../styles.dart';
 
 class HackRUApp extends StatefulWidget {
@@ -23,31 +22,33 @@ class HackRUApp extends StatefulWidget {
 }
 
 class _HackRUAppState extends State<HackRUApp> {
-  List<ScreenHiddenDrawer> items = new List();
+  List<ScreenHiddenDrawer> items = [];
   bool _hasAuthToken = false;
   User user;
 
-  TextStyle _selectedDrawerItem = TextStyle(color: pink, fontWeight: FontWeight.w700);
-  TextStyle _nonSelectedDrawerItem = TextStyle(color: grey, fontSize: 28.0, fontWeight: FontWeight.w500);
+  final _selectedDrawerItem =
+      TextStyle(color: pink, fontWeight: FontWeight.w700);
+  final _nonSelectedDrawerItem =
+      TextStyle(color: grey, fontSize: 28.0, fontWeight: FontWeight.w500);
 
   @override
   void initState() {
     _hasToken();
-    if(_hasAuthToken != null) _drawerItems();
+    if (_hasAuthToken != null) _drawerItems();
     super.initState();
   }
 
   /// =========================================================
   ///                     CREDENTIAL MANAGER
   /// =========================================================
-  void _hasToken() async{
+  void _hasToken() async {
     var hasToken = await hasCredentials();
-    if(hasToken){
+    if (hasToken) {
       setState(() {
         _hasAuthToken = hasToken;
       });
       _getUserProfile();
-    } else{
+    } else {
       setState(() {
         _hasAuthToken = false;
       });
@@ -59,7 +60,7 @@ class _HackRUAppState extends State<HackRUApp> {
     var _authToken = await getAuthToken();
     var userProfile = await getUser(_authToken, _storedEmail);
     print('====== Email: ${userProfile.email} ======');
-    if(userProfile != null){
+    if (userProfile != null) {
       setState(() {
         user = userProfile;
       });
@@ -69,11 +70,11 @@ class _HackRUAppState extends State<HackRUApp> {
   /// =========================================================
   ///                      DRAWER ITEMS
   /// =========================================================
-  void _drawerItems(){
+  void _drawerItems() {
     setState(() {});
-    items.add(new ScreenHiddenDrawer(
-      new ItemHiddenMenu(
-        name: "Home",
+    items.add(ScreenHiddenDrawer(
+      ItemHiddenMenu(
+        name: 'Home',
         baseStyle: _nonSelectedDrawerItem,
         colorLineSelected: yellow,
         selectedStyle: _selectedDrawerItem,
@@ -81,9 +82,9 @@ class _HackRUAppState extends State<HackRUApp> {
       Home(),
     ));
 
-    items.add(new ScreenHiddenDrawer(
-      new ItemHiddenMenu(
-        name: "Map",
+    items.add(ScreenHiddenDrawer(
+      ItemHiddenMenu(
+        name: 'Map',
         baseStyle: _nonSelectedDrawerItem,
         colorLineSelected: yellow,
         selectedStyle: _selectedDrawerItem,
@@ -91,9 +92,9 @@ class _HackRUAppState extends State<HackRUApp> {
       HackRUMap(),
     ));
 
-    items.add(new ScreenHiddenDrawer(
-      new ItemHiddenMenu(
-        name: "Help",
+    items.add(ScreenHiddenDrawer(
+      ItemHiddenMenu(
+        name: 'Help',
         baseStyle: _nonSelectedDrawerItem,
         colorLineSelected: yellow,
         selectedStyle: _selectedDrawerItem,
@@ -101,9 +102,9 @@ class _HackRUAppState extends State<HackRUApp> {
       Help(),
     ));
 
-    items.add(new ScreenHiddenDrawer(
-      new ItemHiddenMenu(
-        name: "About",
+    items.add(ScreenHiddenDrawer(
+      ItemHiddenMenu(
+        name: 'About',
         baseStyle: _nonSelectedDrawerItem,
         colorLineSelected: yellow,
         selectedStyle: _selectedDrawerItem,
@@ -111,9 +112,9 @@ class _HackRUAppState extends State<HackRUApp> {
       About(),
     ));
 
-    items.add(new ScreenHiddenDrawer(
-      new ItemHiddenMenu(
-        name: "QR Scanner",
+    items.add(ScreenHiddenDrawer(
+      ItemHiddenMenu(
+        name: 'QR Scanner',
         baseStyle: _nonSelectedDrawerItem,
         colorLineSelected: yellow,
         selectedStyle: _selectedDrawerItem,
@@ -139,18 +140,19 @@ class _HackRUAppState extends State<HackRUApp> {
 //    }
 
     if (_hasAuthToken) {
-      items.add(new ScreenHiddenDrawer(
-        new ItemHiddenMenu(
-          name: "Logout",
+      items.add(ScreenHiddenDrawer(
+        ItemHiddenMenu(
+          name: 'Logout',
           baseStyle: _nonSelectedDrawerItem,
           colorLineSelected: yellow,
           selectedStyle: _selectedDrawerItem,
           onTap: () async {
             await deleteCredentials();
-            Navigator.of(context).pushAndRemoveUntil(new MaterialPageRoute(
-              builder: (BuildContext context) => HackRUApp(),
-              maintainState: false,
-            ),
+            await Navigator.of(context).pushAndRemoveUntil(
+              MaterialPageRoute(
+                builder: (BuildContext context) => HackRUApp(),
+                maintainState: false,
+              ),
               ModalRoute.withName('/main'),
             );
           },
@@ -158,7 +160,6 @@ class _HackRUAppState extends State<HackRUApp> {
         Container(),
       ));
     }
-
   }
 
   /// =========================================================
@@ -167,35 +168,38 @@ class _HackRUAppState extends State<HackRUApp> {
 
   @override
   Widget build(BuildContext context) {
-    print('==**==== $_hasAuthToken ===**===');
+    print('==**==== HAS_AUTH_TOKEN: $_hasAuthToken ===**===');
     return CustomHiddenDrawerMenu(
       actionsAppBar: <Widget>[
-        _hasAuthToken ? IconButton(
-          icon: Icon(
-            GroovinMaterialIcons.logout,
-            color: Theme.of(context).primaryColor,
-          ),
-          color: transparent,
-          splashColor: yellow,
-          onPressed: () async {
-            await deleteCredentials();
-            setState(() {
-              _hasAuthToken = false;
-            });
-            Navigator.of(context).pushAndRemoveUntil(new MaterialPageRoute(
-              builder: (BuildContext context) => HackRUApp(),
-              maintainState: false,
-            ),
-              ModalRoute.withName('/main'),
-            );
-          },
-        ) : Container(),
+        _hasAuthToken
+            ? IconButton(
+                icon: Icon(
+                  GroovinMaterialIcons.logout,
+                  color: Theme.of(context).primaryColor,
+                ),
+                color: transparent,
+                splashColor: yellow,
+                onPressed: () async {
+                  await deleteCredentials();
+                  setState(() {
+                    _hasAuthToken = false;
+                  });
+                  await Navigator.of(context).pushAndRemoveUntil(
+                    MaterialPageRoute(
+                      builder: (BuildContext context) => HackRUApp(),
+                      maintainState: false,
+                    ),
+                    ModalRoute.withName('/main'),
+                  );
+                },
+              )
+            : Container(),
       ],
-      iconMenuAppBar: Icon(
+      leadingAppBar: Icon(
         Icons.menu,
         color: grey,
       ),
-      styleAutoTittleName: Theme.of(context).textTheme.title,
+      styleAutoTittleName: Theme.of(context).textTheme.headline6,
       backgroundColorMenu: grey,
       backgroundColorAppBar: Theme.of(context).backgroundColor,
       elevationAppBar: 0.0,
@@ -205,7 +209,7 @@ class _HackRUAppState extends State<HackRUApp> {
           'assets/flare/party.flr',
           alignment: Alignment.center,
           fit: BoxFit.contain,
-          animation: "idle",
+          animation: 'idle',
         ),
       ),
       screens: items,
