@@ -1,16 +1,11 @@
-import 'dart:async';
-
-import 'package:HackRU/models/cred_manager.dart';
-import 'package:HackRU/styles.dart';
-import 'package:HackRU/services/hackru_service.dart';
 import 'package:HackRU/models/models.dart';
 import 'package:HackRU/models/string_parser.dart';
-import 'package:flare_flutter/flare_actor.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import 'package:webview_flutter/webview_flutter.dart';
+
+import '../../../styles.dart';
 
 class AnnouncementCard extends StatelessWidget {
   AnnouncementCard({@required this.resource});
@@ -20,9 +15,9 @@ class AnnouncementCard extends StatelessWidget {
   ///               CHECK FOR WEB-LINK
   /// =================================================
 //  bool _isWebLink(String input) {
-//    final matcher = new RegExp(
-//        r"(http(s)?:\/\/.)?(www\.)?[-a-zA-Z0-9@:%._\+~#=]{2,256}\.[a-z]{2,6}\b([-a-zA-Z0-9@:%_\+.~#?&//=]*)");
-//    final specialChar = new RegExp("/");
+//    final matcher =  RegExp(
+//        r'(http(s)?:\/\/.)?(www\.)?[-a-zA-Z0-9@:%._\+~#=]{2,256}\.[a-z]{2,6}\b([-a-zA-Z0-9@:%_\+.~#?&//=]*)');
+//    final specialChar =  RegExp('/');
 //    return matcher.hasMatch(input) && specialChar.hasMatch(input);
 //  }
 
@@ -34,8 +29,8 @@ class AnnouncementCard extends StatelessWidget {
 //    WebView webView;
 //    words.forEach((link) {
 //      if (_isWebLink(link)) {
-//        var linkText = link.replaceAll(new RegExp(r'[<>]'), '');
-//        webView = new WebView(
+//        var linkText = link.replaceAll( RegExp(r'[<>]'), '');
+//        webView =  WebView(
 //          initialUrl: linkText ?? '[null]',
 //          gestureRecognizers: [
 //            Factory(() => PlatformViewVerticalGestureRecognizer()),
@@ -47,17 +42,21 @@ class AnnouncementCard extends StatelessWidget {
 //    return webView;
 //  }
 
+  @override
   Widget build(BuildContext context) {
-    String secs = resource.ts.split(".")[0];
+    final _errorLoadingDataText = 'Error: Unable to retrieve messages!';
+    var secs = resource.ts.split('.')[0];
     var timeStr =
         DateTime.fromMillisecondsSinceEpoch(int.parse(secs) * 1000).toLocal();
-    var formattedTime = new DateFormat('hh:mm a').format(timeStr);
+    var formattedTime = DateFormat('hh:mm a').format(timeStr);
 
-    return new Container(
+    return Container(
       key: Key(resource.ts),
-      child: new Card(
+      child: Card(
         elevation: 0.0,
-        color: Theme.of(context).dividerColor,
+        color: resource.text == _errorLoadingDataText
+            ? pink
+            : Theme.of(context).dividerColor,
         shape: RoundedRectangleBorder(
           borderRadius: BorderRadius.circular(15.0),
         ),
@@ -65,22 +64,37 @@ class AnnouncementCard extends StatelessWidget {
           padding: const EdgeInsets.all(10.0),
           child: Column(
             children: <Widget>[
-              ListTile(
-                title: Text(
-                  formattedTime,
-                  style: TextStyle(fontSize: 18.0, color: Theme.of(context).primaryColorDark, fontWeight: FontWeight.w700,),
-                ),
-                subtitle: StringParser(
-                  text: resource.text ?? '',
-                ),
-              ),
+              resource.text == _errorLoadingDataText
+                  ? Center(
+                      child: Text(
+                        _errorLoadingDataText,
+                        style: Theme.of(context)
+                            .textTheme
+                            .subtitle1
+                            .copyWith(color: Colors.white),
+                      ),
+                    )
+                  : ListTile(
+                      title: Text(
+                        formattedTime,
+                        style: TextStyle(
+                          fontSize: 18.0,
+                          color: Theme.of(context).primaryColorDark,
+                          fontWeight: FontWeight.w700,
+                        ),
+                      ),
+                      subtitle: StringParser(
+                        text: resource.text ?? '',
+                      ),
+                    ),
+
               /// TODO: Decide whether to show WebView or only meta image of a website
-//            new Card(
+//             Card(
 //              elevation: 0.0,
 //              shape: RoundedRectangleBorder(
 //                borderRadius: BorderRadius.circular(15.0),
 //              ),
-//              child: new Container(
+//              child:  Container(
 //                child: _enableWebview(resource.text),
 //                height: _isWebLink(resource.text) ? 250.0 : 0.0,
 //              ),
@@ -114,8 +128,8 @@ class PlatformViewVerticalGestureRecognizer
   void handleEvent(PointerEvent event) {
     _dragDistance = _dragDistance + event.delta;
     if (event is PointerMoveEvent) {
-      final double dy = _dragDistance.dy.abs();
-      final double dx = _dragDistance.dx.abs();
+      final dy = _dragDistance.dy.abs();
+      final dx = _dragDistance.dx.abs();
 
       if (dy > dx && dy > kTouchSlop) {
         // vertical drag - accept
