@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 /// Reference: https://github.com/flutter/flutter/blob/2d2a1ffec9/packages/flutter/lib/src/material/expansion_tile.dart#L30
 
-const Duration _kExpand = const Duration(milliseconds: 200);
+const Duration _kExpand = Duration(milliseconds: 200);
 
 class CustomExpansionTile extends StatefulWidget {
   const CustomExpansionTile({
@@ -11,9 +11,9 @@ class CustomExpansionTile extends StatefulWidget {
     @required this.title,
     this.backgroundColor,
     this.onExpansionChanged,
-    this.children: const <Widget>[],
+    this.children = const <Widget>[],
     this.trailing,
-    this.initiallyExpanded: false,
+    this.initiallyExpanded = false,
   })  : assert(initiallyExpanded != null),
         super(key: key);
 
@@ -26,7 +26,7 @@ class CustomExpansionTile extends StatefulWidget {
   final bool initiallyExpanded;
 
   @override
-  CustomExpansionTileState createState() => new CustomExpansionTileState();
+  CustomExpansionTileState createState() => CustomExpansionTileState();
 }
 
 class CustomExpansionTileState extends State<CustomExpansionTile>
@@ -45,17 +45,16 @@ class CustomExpansionTileState extends State<CustomExpansionTile>
   @override
   void initState() {
     super.initState();
-    _controller = new AnimationController(duration: _kExpand, vsync: this);
+    _controller = AnimationController(duration: _kExpand, vsync: this);
     _easeOutAnimation =
-        new CurvedAnimation(parent: _controller, curve: Curves.easeOut);
+        CurvedAnimation(parent: _controller, curve: Curves.easeOut);
     _easeInAnimation =
-        new CurvedAnimation(parent: _controller, curve: Curves.easeIn);
-    _borderColor = new ColorTween();
-    _headerColor = new ColorTween();
-    _iconColor = new ColorTween();
-    _iconTurns =
-        new Tween<double>(begin: 0.0, end: 0.5).animate(_easeInAnimation);
-    _backgroundColor = new ColorTween();
+        CurvedAnimation(parent: _controller, curve: Curves.easeIn);
+    _borderColor = ColorTween();
+    _headerColor = ColorTween();
+    _iconColor = ColorTween();
+    _iconTurns = Tween<double>(begin: 0.0, end: 0.5).animate(_easeInAnimation);
+    _backgroundColor = ColorTween();
 
     _isExpanded =
         PageStorage.of(context)?.readState(context) ?? widget.initiallyExpanded;
@@ -84,15 +83,16 @@ class CustomExpansionTileState extends State<CustomExpansionTile>
     if (_isExpanded != isExpanded) {
       setState(() {
         _isExpanded = isExpanded;
-        if (_isExpanded)
+        if (_isExpanded) {
           _controller.forward();
-        else
+        } else {
           _controller.reverse().then<void>((void value) {
             if (!mounted) return;
             setState(() {
               // Rebuild without widget.children.
             });
           });
+        }
         PageStorage.of(context)?.writeState(context, _isExpanded);
       });
       if (widget.onExpansionChanged != null) {
@@ -102,43 +102,42 @@ class CustomExpansionTileState extends State<CustomExpansionTile>
   }
 
   Widget _buildChildren(BuildContext context, Widget child) {
-    final Color borderSideColor =
+    final borderSideColor =
         _borderColor.evaluate(_easeOutAnimation) ?? Colors.transparent;
-    final Color titleColor = _headerColor.evaluate(_easeInAnimation);
+    final titleColor = _headerColor.evaluate(_easeInAnimation);
 
-    return new Container(
-      decoration: new BoxDecoration(
+    return Container(
+      decoration: BoxDecoration(
           color: _backgroundColor.evaluate(_easeOutAnimation) ??
               Colors.transparent,
-          border: new Border(
-            top: new BorderSide(color: borderSideColor),
-            bottom: new BorderSide(color: borderSideColor),
+          border: Border(
+            top: BorderSide(color: borderSideColor),
+            bottom: BorderSide(color: borderSideColor),
           )),
-      child: new Column(
+      child: Column(
         mainAxisSize: MainAxisSize.min,
         children: <Widget>[
           IconTheme.merge(
-            data:
-                new IconThemeData(color: _iconColor.evaluate(_easeInAnimation)),
-            child: new ListTile(
+            data: IconThemeData(color: _iconColor.evaluate(_easeInAnimation)),
+            child: ListTile(
               onTap: toggle,
               leading: widget.leading,
-              title: new DefaultTextStyle(
+              title: DefaultTextStyle(
                 style: Theme.of(context)
                     .textTheme
-                    .subhead
+                    .subtitle1
                     .copyWith(color: titleColor),
                 child: widget.title,
               ),
               trailing: widget.trailing ??
-                  new RotationTransition(
+                  RotationTransition(
                     turns: _iconTurns,
                     child: const Icon(Icons.expand_more),
                   ),
             ),
           ),
-          new ClipRect(
-            child: new Align(
+          ClipRect(
+            child: Align(
               heightFactor: _easeInAnimation.value,
               child: child,
             ),
@@ -150,21 +149,21 @@ class CustomExpansionTileState extends State<CustomExpansionTile>
 
   @override
   Widget build(BuildContext context) {
-    final ThemeData theme = Theme.of(context);
+    final theme = Theme.of(context);
     _borderColor.end = theme.dividerColor;
     _headerColor
-      ..begin = theme.textTheme.subhead.color
+      ..begin = theme.textTheme.subtitle1.color
       ..end = theme.accentColor;
     _iconColor
       ..begin = theme.unselectedWidgetColor
       ..end = theme.accentColor;
     _backgroundColor.end = widget.backgroundColor;
 
-    final bool closed = !_isExpanded && _controller.isDismissed;
-    return new AnimatedBuilder(
+    final closed = !_isExpanded && _controller.isDismissed;
+    return AnimatedBuilder(
       animation: _controller.view,
       builder: _buildChildren,
-      child: closed ? null : new Column(children: widget.children),
+      child: closed ? null : Column(children: widget.children),
     );
   }
 }
